Refactor evaluation components
