+++ src/app/evaluations/FieldsCard.tsx
@@ -2,7 +2,6 @@
 
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { useQuery } from "convex/react";
-import { listDynamicFieldsByCreator } from "../../../convex/dynamic_fields";
 import { api } from "../../../convex/_generated/api";
 import { Card } from "@/components/ui/card";
 import { useSession } from "@clerk/nextjs";
